chore: rename module name from graphql to gql_models to resolve import confusion
